+++ EDCA.ps1
@@ -2,7 +2,7 @@
 .SYNOPSIS
     EDCA — Exchange Deployment & Compliance Assessment.
 
-    Version: 0.6 Preview
+    Version: 0.7 Preview
     Author:  Michel de Rooij
     Source:  https://github.com/michelderooij/EDCA
     Website: https://eightwone.com
@@ -127,7 +127,7 @@ param(
 Set-StrictMode -Version Latest
 $ErrorActionPreference = 'Stop'
 
-$EDCAVersion = 'v0.6 Preview'
+$EDCAVersion = 'v0.7 Preview'
 
 $scriptRoot = Split-Path -Parent $MyInvocation.MyCommand.Path